refactored unit testing
